requirements.txt
django==4.2
djangorestframework
cloudinary
python-decouple
djangorestframework-simplejwt
drf-yasg
django-cors-headers
psycopg2
ckeditor